chore: 폴더 정리
