+++ apps/src/app/login/page.tsx
@@ -1,4 +1,4 @@
-import { Login } from '@/components/Login';
+import { Login } from "@/features/auth/ui/Login";
 
 export default function LoginPage() {
   return (
