+++ apps/src/app/profile/page.tsx
@@ -1,4 +1,4 @@
-import UserProfileTemp from "@/components/UserProfile-temp";
+import UserProfileTemp from "@/entities/user/ui/UserProfile-temp";
 
 export default function ProfilePage() {
     return <UserProfileTemp />;
