+++ apps/src/features/home/ui/CategorySection.tsx
