+++ apps/src/features/home/ui/NoticeSection.tsx
