+++ apps/src/widgets/chatList/ui/ChatListScreen.tsx
@@ -1,7 +1,6 @@
 import StatusBar from '../../../features/chat/ui/StatusBar';
 import ChatListHeader from './ChatListHeader';
 import ChatList from './ChatList';
-import BottomBanner from './BottomBanner';
 import HomeIndicator from '../../../features/chat/ui/HomeIndicator';
 import Link from 'next/link';
 
@@ -11,7 +10,6 @@ export default function ChatListScreen() {
       <StatusBar />
       <ChatListHeader />
       <ChatList />
-      <BottomBanner />
       <HomeIndicator />
     </Link>
   );
